fix ArrowHead numbers exceeding int32 mappings (#104)
* fix ArrowHead numbers exceeding int32 mappings

* sneaking in a dev key fix

+++ src/Helldivers-2-Models/ArrowHead/Status/Campaign.cs
@@ -11,5 +11,5 @@ public sealed record Campaign(
     int Id,
     int PlanetIndex,
     int Type,
-    int Count
+    ulong Count
 );
